refactor: tidy up pattern matching
